requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.12.0
aniso8601==9.0.1
annotated-types==0.5.0
anyio==3.7.1
apache-airflow==2.7.1
apache-airflow-providers-common-sql==1.7.2
apache-airflow-providers-ftp==3.5.2
apache-airflow-providers-http==4.5.2
apache-airflow-providers-imap==3.3.2
apache-airflow-providers-sqlite==3.4.3
apispec==6.3.0
argcomplete==3.1.2
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.6.2
bs4==0.0.1
cachelib==0.9.0
cachetools==5.3.1
cattrs==23.1.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
chroma-hnswlib==0.7.3
chromadb
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
coloredlogs==15.0.1
colorlog==4.8.0
ConfigUpdater==3.1.1
connexion==2.14.2
cron-descriptor==1.4.0
croniter==1.4.1
cryptography==41.0.4
dataclasses-json==0.5.14
Deprecated==1.2.14
dill==0.3.7
dnspython==2.4.2
docutils==0.20.1
email-validator==1.3.1
exceptiongroup==1.1.3
fastapi==0.99.1
filelock==3.12.4
Flask==2.2.5
Flask-AppBuilder==4.3.6
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.5.0
Flask-Login==0.6.2
Flask-RESTful==0.3.10
Flask-Session==0.5.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.9.1
google-api-core==2.11.1
google-auth==2.23.0
google-cloud==0.34.0
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-cloud-pubsub==2.18.4
google-cloud-pubsublite
google-crc32c==1.5.0
google-re2==1.1
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
graphviz==0.20.1
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.58.0
grpcio-status==1.58.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
inflection==0.5.1
interchange==2021.0.4
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
langchain==0.0.294
langsmith==0.0.38
lazy-object-proxy==1.9.0
limits==3.6.0
linkify-it-py==2.0.2
lockfile==0.12.2
lxml==4.9.3
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.6
numpy==1.26.0
onnxruntime==1.15.1
openai==0.28.0
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-exporter-otlp-proto-http==1.20.0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
ordered-set==4.1.0
outcome==1.2.0
overrides
packaging==23.1
pansi==2020.7.3
pathspec==0.11.2
pendulum==2.1.2
pluggy==1.3.0
posthog==3.0.2
prison==0.2.1
proto-plus==1.22.3
protobuf==4.24.3
psutil==5.9.5
pulsar-client==3.3.0
py2neo==2021.2.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.6.3
Pygments==2.16.1
PyJWT==2.8.0
pymongo==4.5.0
pyOpenSSL==23.2.0
pypdf==3.16.1
PyPika==0.48.9
PySocks==1.7.1
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.7.1
python-nvd3==0.15.0
python-slugify==8.0.1
python-socketio==5.9.0
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.5.3
rich-argparse==1.3.0
rpds-py==0.10.3
rsa==4.9
selenium==4.12.0
setproctitle==1.3.2
simple-websocket==0.10.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==1.4.49
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
text-unidecode==1.3
tiktoken==0.5.1
tokenizers==0.14.0
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.10.4
typing-inspect==0.9.0
typing_extensions==4.7.1
uc-micro-py==1.0.2
unicodecsv==0.14.1
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
Werkzeug==2.2.3
wrapt==1.15.0
wsproto==1.2.0
WTForms==3.0.1
yarl==1.9.2
zipp==3.16.2
